+++ src/runtime/utils/string_util.h
@@ -12,9 +12,9 @@ namespace utils
 {
 
 #define DUP_STR_TO_LOCAL_TEMP_ZERO_END_STR(local_temp_str, str, str_len) \
-    char* local_temp_str = (char*)alloca(str_len + 1);                   \
-    std::memcpy(local_temp_str, str, str_len);                           \
-    local_temp_str[str_len] = '\0';
+    char* local_temp_str = (char*)alloca(static_cast<size_t>(str_len) + 1); \
+    std::memcpy(local_temp_str, str, static_cast<size_t>(str_len));        \
+    local_temp_str[static_cast<size_t>(str_len)] = '\0';
 
 class StringBuilder;
